graphics and structure

+++ Agar.io/Agar.IO.Client.Console/Models/Position.cs
@@ -6,10 +6,10 @@ namespace Agar.IO.Client.WinForms.Models
 {
     struct Position
     {
-        public float X;
-        public float Y;
+        public double X;
+        public double Y;
 
-        public Position(float x, float y)
+        public Position(double x, double y)
         {
             X = x;
             Y = y;
